Restructuring configuration for module celebrate
